+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Added
+- `cucumberStepStart` and `cucumberStepEnd` commands for reporting `cypress-cucumber-preprocessor` scenario steps as nested steps in RP.
+### Security
+- Updated versions of vulnerable packages (@reportportal/client-javascript, glob).
 
 ## [5.2.0] - 2024-03-21
 ### Fixed